% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{chol2cor}
\alias{chol2cor}
\title{Create a correlation matrix from Cholesky decomposition}
\usage{
chol2cor(vec, dime)
}
\arguments{
\item{vec}{vector of parameters for the Cholesky decomposition
(upper diagonal matrix, excluding the diagonal)}

\item{dime}{dimension of the model}
}
\value{
a correlation matrix
}
\description{
This function takes a vector, fills a Cholesky upper triangular matrix and transforms it to a
positive definite matrix, then adds a diagonal of ones for the extremal Student model
}
\keyword{internal}

