% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{chol2Lambda}
\alias{chol2Lambda}
\title{Create a conditionally negative definite matrix from Cholesky decomposition}
\usage{
chol2Lambda(vec, dime, EoW = TRUE)
}
\arguments{
\item{vec}{vector of parameters for the Cholesky decomposition, with log-diagonal entries}

\item{dime}{dimension of the model}
}
\value{
a CNSD matrix
}
\description{
This function takes a vector, fills a Cholesky upper triangular matrix and transforms it to a
positive definite matrix, then into a symmetric CNSD matrix with zero diagonal for the HR model
}
\keyword{internal}

