% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{composite_pot_startvals}
\alias{composite_pot_startvals}
\title{Starting values for multivariate peaks-over-threshold via pairwise composite likelihood methods}
\usage{
composite_pot_startvals(x, u = NULL, model, sym = FALSE, cscale = FALSE,
  cshape = FALSE, transform = FALSE)
}
\arguments{
\item{x}{\code{n} by \code{d} data matrix on original scale}

\item{u}{\code{d} vector of threshold parameters on probability scale}

\item{model}{string indicating the multivariate extreme value family}

\item{sym}{logical indicating whether model should be made symmetric}

\item{cscale}{logical indicating common scale for all variables}

\item{cshape}{logical indicating common shape for all variables}

\item{transform}{logical indicating transformation to unconstrained domain for parameters}
}
\description{
This function is the analog of \link[evd]{bvstart.vals}, except that it handles higher dimensional models
}
\keyword{internal}

