% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{compositemat}
\alias{compositemat}
\title{Estimation of the variability and Godambe information matrix using a
nonparametric bootstrap scheme}
\usage{
compositemat(dat, fitted, B, use.start = FALSE, ...)
}
\arguments{
\item{dat}{data matrix}

\item{fitted}{output from the call to \code{fmvcpot}.}

\item{B}{number of bootstrap replicates}

\item{use.start}{logical indicating whether to use MCLE as starting value}

\item{...}{fixed pararameters to pass to \code{fmvcpot} if any}
}
\value{
a list with matrices \code{godambe}, \code{sensitivity} and \code{variability}
}
\description{
Composite likelihood estimators are consistent and asymptotically normal
under mild regularity conditions. The asymptotic covariance matrix is the
inverse Godambe information matrix \eqn{G^{-1}}, which is estimated by a
nonparametric bootstrap as the empirical covariance of \code{B} replicates.
The sensitivity matrix \eqn{H} is calculated from the Hessian matrix at the
maximum composite likelihood estimates. Lastly, the variability matrix \eqn{J} is
obtained from the relation \eqn{G=HJ^{-1}H}
}

