% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientscore.R
\name{fscore}
\alias{fscore}
\title{Fit Dirichlet extremal model using the gradient score}
\usage{
fscore(start, dat, model, p = 1)
}
\arguments{
\item{start}{starting values for the parameter (\eqn{alpha} and \eqn{rho}, in this order)}

\item{dat}{matrix of transformed data on unit Pareto scale lying above the threshold}

\item{model}{string indicating whether the \code{"dir"} or \code{"negdir"} model should be considered}

\item{p}{int specifying the degree of the \eqn{l_p} norm used in the weight function \code{weightFun}}
}
\description{
Optimization routine to fit a member of the extremal Dirichlet family
using the gradient score
}

