% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isAbove}
\alias{isAbove}
\title{Indicates whether observations lie above the threshold.}
\usage{
isAbove(data, threshold)
}
\arguments{
\item{data}{matrix of all observations}

\item{threshold}{vector of thresholds under which observations are censored}
}
\value{
matrix of logical indicating whether or not observations lie above the marginal threshold
}
\description{
The function returns a boolean matrix of indicators.
}
\keyword{internal}

