% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{Lambda2Sigma}
\alias{Lambda2Sigma}
\title{Transform a symmetric conditionally negative definite matrix into a covariance matrix}
\usage{
Lambda2Sigma(Lambda)
}
\arguments{
\item{Lambda}{symmetric conditionally negative definite matrix of parameters for the Husler-Reiss model}
}
\value{
a symmetric positive definite matrix
}
\description{
Transform a symmetric conditionally negative definite matrix into a covariance matrix
}
\keyword{internal}

