% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MarginalTransformation}
\alias{MarginalTransformation}
\title{Marginal transformation of the observations to Frechet or Pareto scale}
\usage{
MarginalTransformation(data, u, lambda, scale, shape, GEVorGP = TRUE)
}
\arguments{
\item{data}{\code{n} by \code{d} matrix of observations on original scale}

\item{u}{threshold (on \code{y}-scale), to be substracted from y}

\item{lambda}{vector of percentage of threshold exceedances}

\item{scale}{vector of scale parameters}

\item{shape}{vector of shape parameters}

\item{GEVorGP}{logical indicating whether to transform to unit Frechet or unit Pareto. Default to \code{true} (GEV).}
}
\description{
This function uses the probability integral transform to map the observations
to unit Frechet scale \code{GEV(1,1,1)} or unit Pareto distribution \code{GP(1,1,1)}.
The function returns a list with the transformed observation and the marginal transformation
Jacobian.
}

