% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nllmvsdir}
\alias{nllmvsdir}
\title{Negative log-likelihood of scaled Dirichlet extreme value model}
\usage{
nllmvsdir(y, thid, N, lambda, u, alpha, rho, scale, shape)
}
\arguments{
\item{y}{\code{n} by \code{d} matrix of observations on original scale}

\item{thid}{\code{n} by \code{d} matrix of logical vector indicating whether or not a
threshold exceedance was observed}

\item{N}{total number of observations (not just exceedances)}

\item{lambda}{vector of percentage of threshold exceedances}

\item{u}{threshold (on \code{y}-scale), to be substracted from y}

\item{alpha}{vector of parameters for the Dirichlet model}

\item{rho}{regular variation parameter}

\item{scale}{vector of scale parameters}

\item{shape}{vector of shape parameters}
}
\value{
the negative log-composite likelihood value
}
\description{
This is the workhorse with the likelihood, which transforms observations
and returns the objective function value.
}
\section{Note}{
 The \code{N} argument is there mostly for cases in which function returns
observations stripped from their contribution; in such cases, the count
would otherwise be `incorrect'.
}

