% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{nlloptnegdir}
\alias{nlloptnegdir}
\title{Wrapper for the scaled negative extremal Dirichlet model}
\usage{
nlloptnegdir(optpar, fixedpar, wfixed, numpar, transform = FALSE, dat, thid,
  N, lambda, u, datmarg)
}
\arguments{
\item{optpar}{vector of parameters over which to optimize}

\item{fixedpar}{vector of parameters to keep constant}

\item{wfixed}{is vector of int with indices of data points}

\item{numpar}{vector integer giving the length of each parameter (scale, shape, ...)}

\item{transform}{boolean indicating whether to exponentiate the parameters for optimization}

\item{dat}{data matrix}

\item{thid}{logical matrix indicating whether entries exceed the marginal threshold}

\item{N}{total sample size}

\item{lambda}{vector of marginal rates of exceedance}

\item{u}{vector of thresholds}
}
\value{
the negative log-likelihood value
}
\description{
This function takes as argument the parameters of the scaled Dirichlet bivariate distribution
and passes them to the C++ function.
}
\keyword{internal}

