% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{Sigma2Lambda}
\alias{Sigma2Lambda}
\title{Transform covariance matrix to conditionally negative definite matrix}
\usage{
Sigma2Lambda(Sigma)
}
\arguments{
\item{Sigma}{symmetric positive definite matrix}
}
\value{
a CND matrix
}
\description{
Transform covariance matrix to conditionally negative definite matrix
}
\keyword{internal}

