% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraldensities.R
\name{specdens}
\alias{specdens}
\title{Spectral density of scaled Dirichlet and negative Dirichlet extreme value distributions}
\usage{
specdens(param, dat, d, model = c("dir", "negdir", "ct"), transform = TRUE)
}
\arguments{
\item{param}{vector containing the parameters \eqn{\alpha} and \eqn{\rho}, in this order.}

\item{dat}{matrix of Fréchet or Pareto observations, of dimension \eqn{n} by \eqn{d}.}

\item{d}{dimension of model}

\item{model}{string indicating which model to select from \code{"dir"}, \code{"negdir"} and \code{"ct"}}

\item{transform}{logical indicating whether parameters are on the log scale. Default to \code{TRUE}}
}
\value{
the log-density for the \code{n} sample
}
\description{
Spectral density of scaled Dirichlet and negative Dirichlet extreme value distributions
}
\details{
The function is provided as a wrapper and takes parameters on the log scale for \eqn{\alpha} (\eqn{\rho}).
}

