% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcomposite.R
\name{fmvcpot}
\alias{fmvcpot}
\title{Pairwise composite likelihood estimation for peaks-over-threshold}
\usage{
fmvcpot(dat, u, lambda, N, model = c("ct", "dir", "negdir", "hr"),
  cscale = FALSE, cshape = FALSE, sym = FALSE, start = NULL,
  std.err = TRUE, method = "BFGS", warn.inf = TRUE, ...)
}
\arguments{
\item{dat}{\code{n} by \code{d} data matrix on original scale}

\item{u}{\code{d} vector of threshold parameters on probability scale}

\item{lambda}{vector of percentage of threshold exceedances (can be omitted)}

\item{N}{total number of observations if some observations falling under all threshold were removed from \code{x} (can be omitted)}

\item{model}{family of multivariate extreme value distribution. \code{'ct'} and \code{'hr'} are the only family currently supported.}

\item{cscale}{logical indicating common scale for all variables}

\item{cshape}{logical indicating common shape for all variables}

\item{sym}{logical indicating whether equal dependence parameters in the Dirichlet model}

\item{start}{vector of starting value for the optimization}

\item{std.err}{logical indicating whether to calculate standard errors numerically}

\item{method}{optimization method; see \link[stats]{optim}}

\item{warn.inf}{whether to check the validity of starting values before optimization}

\item{...}{additional arguments (fixed parameters) passed to the \code{optim}
function to be held fixed during the optimization}

\item{transform}{logical indicating whether to perform optimization on unconstrained
space; \code{start} should be adjusted if \code{transform=TRUE}}
}
\value{
a list containing the output of \code{\link[stats]{optim}}, notably
parameter values \code{par} and standard errors \code{se}
}
\description{
This function finds the maximum composite likelihood estimate
for multivariate extreme value models based on pairs of data.
It returns a list with the result from \code{\link[stats]{optim}}
along with some of the information passed to the function.
}
\examples{
x <- rmev(n=100000, d=2, param=parame, model = "negdir")
qu <- apply(x, 2, quantile,probs <- 0.99)
y <- x[which(rowSums(isAbove(x,threshold = qu))>=1),]
fit <- fmvcpot(dat=x, u=qu,model="negdir",lambda=colSums(isAbove(y,qu))/(nrow(x)+1), N=nrow(x),
               cscale=TRUE, cshape=TRUE, shape=1, scale=100)
fit$par
}

